update: todas las operaciones necesitan un  jwt

+++ controllers/itemController.js
@@ -6,98 +6,126 @@ const secret = "secret"; // Debe ser la misma clave utilizada al firmar el token
 
 exports.createItem = async (req, res) => {
   /* 
-     #swagger.tags = ['Items']
-     #swagger.description = 'Create an item'
-     #swagger.summary = 'Create an item'
-     #swagger.parameters['data'] = {
-         in: 'body',
-         description: 'Data to create an item',
-         required: true,
-     }
-     #swagger.responses[201] = {
-         description: 'Item successfully created',
-     }
-     #swagger.responses[400] = {
-         description: 'Bad request',
-     }
-   */
+    #swagger.tags = ['Items']
+    #swagger.description = 'Create an item'
+    #swagger.summary = 'Create an item'
+    #swagger.parameters['data'] = {
+      in: 'body',
+      description: 'Data to create an item',
+      required: true,
+    }
+    #swagger.parameters['token'] = {
+      in: 'header',
+      description: 'JWT Token',
+      required: true,
+    }
+    #swagger.responses[201] = {
+      description: 'Item successfully created',
+    }
+    #swagger.responses[400] = {
+      description: 'Bad request',
+    }
+  */
 
-  try {
-    const data = req.body;
-    const itemRef = await db.collection("items").add(data);
-    res.status(201).send(`Created a new item: ${itemRef.id}`);
-  } catch (error) {
-    res.status(400).send(error.message);
-  }
-};
-
-exports.getAllItems = async (req, res) => {
-  /* 
-     #swagger.tags = ['Items']
-     #swagger.description = 'Get all items entries'
-     #swagger.summary = 'Get all items entries'
-     #swagger.parameters['token'] = {
-         in: 'header',
-         description: 'JWT Token',
-         required: true,
-     }
-     #swagger.responses[200] = {
-         description: 'Items entries successfully obtained',
-     }
-     #swagger.responses[400] = {
-         description: 'Bad request',
-     }
-   */
-
-  const token = req.headers['token'];
+  const token = req.headers["token"];
   if (!token) {
     return res.status(400).send("Token no proporcionado");
   }
 
   try {
-
     const decoded = jwt.verify(token, secret);
-    console.log("Token valido:", decoded)
+    console.log("Token valido:", decoded);
 
-    const itemsSnapshot = await db.collection("items").get();
-    const items = [];
-    itemsSnapshot.forEach((doc) => items.push({ id: doc.id, ...doc.data() }));
-    res.status(200).json(items);
-
+    const data = req.body;
+    const itemRef = await db.collection("items").add(data);
+    res.status(201).send(`Created a new item: ${itemRef.id}`);
   } catch (error) {
-
-    if (error.name === 'TokenExpiredError') {
+    if (error.name === "TokenExpiredError") {
       // Token ha expirado
       return res.status(408).send("El token ha expirado");
     } else {
       // Otro error
       return res.status(400).send(error.message);
     }
+  }
+};
 
+exports.getAllItems = async (req, res) => {
+  /* 
+    #swagger.tags = ['Items']
+    #swagger.description = 'Get all items entries'
+    #swagger.summary = 'Get all items entries'
+    #swagger.parameters['token'] = {
+      in: 'header',
+      description: 'JWT Token',
+      required: true,
+    }
+    #swagger.responses[200] = {
+      description: 'Items entries successfully obtained',
+    }
+    #swagger.responses[400] = {
+      description: 'Bad request',
+    }
+  */
+
+  const token = req.headers["token"];
+  if (!token) {
+    return res.status(400).send("Token no proporcionado");
+  }
+
+  try {
+    const decoded = jwt.verify(token, secret);
+    console.log("Token valido:", decoded);
+
+    const itemsSnapshot = await db.collection("items").get();
+    const items = [];
+    itemsSnapshot.forEach((doc) => items.push({ id: doc.id, ...doc.data() }));
+    res.status(200).json(items);
+  } catch (error) {
+    if (error.name === "TokenExpiredError") {
+      // Token ha expirado
+      return res.status(408).send("El token ha expirado");
+    } else {
+      // Otro error
+      return res.status(400).send(error.message);
+    }
   }
 };
 
 exports.getItem = async (req, res) => {
   /* 
-     #swagger.tags = ['Items']
-     #swagger.description = 'Get an item entry'
-     #swagger.summary = 'Get an item entry'
-     #swagger.parameters['id'] = {
-         description: 'Item id',
-         required: true,
-     }
-     #swagger.responses[404] = {
-         description: 'Item not found',
-     }
-     #swagger.responses[400] = {
-         description: 'Bad request',
-     }
-     #swagger.responses[200] = {
-         description: 'Get an item by id',
-     }
-   */
+    #swagger.tags = ['Items']
+    #swagger.description = 'Get an item entry'
+    #swagger.summary = 'Get an item entry'
+    #swagger.parameters['id'] = {
+      description: 'Item id',
+      required: true,
+    }
+    #swagger.parameters['token'] = {
+      in: 'header',
+      description: 'JWT Token',
+      required: true,
+    }
+    #swagger.responses[404] = {
+      description: 'Item not found',
+    }
+    #swagger.responses[400] = {
+      description: 'Bad request',
+    }
+    #swagger.responses[200] = {
+      description: 'Get an item by id',
+    }
+  */
+
+  const token = req.headers["token"];
+  if (!token) {
+    return res.status(400).send("Token no proporcionado");
+  }
 
   try {
+    const decoded = jwt.verify(token, secret);
+    console.log("Token valido:", decoded);
+
     const itemId = req.params.id;
     const itemDoc = await db.collection("items").doc(itemId).get();
     if (!itemDoc.exists) {
@@ -106,7 +134,13 @@ exports.getItem = async (req, res) => {
       res.status(200).json({ id: itemDoc.id, ...itemDoc.data() });
     }
   } catch (error) {
-    res.status(400).send(error.message);
+    if (error.name === "TokenExpiredError") {
+      // Token ha expirado
+      return res.status(408).send("El token ha expirado");
+    } else {
+      // Otro error
+      return res.status(400).send(error.message);
+    }
   }
 };
 
@@ -124,6 +158,11 @@ exports.updateItem = async (req, res) => {
       description: '',
       required: true,
     }
+    #swagger.parameters['token'] = {
+      in: 'header',
+      description: 'JWT Token',
+      required: true,
+    }
     #swagger.responses[200] = {
       description: '',
     }
@@ -132,14 +171,28 @@ exports.updateItem = async (req, res) => {
     }
   */
 
+  const token = req.headers["token"];
+  if (!token) {
+    return res.status(400).send("Token no proporcionado");
+  }
+
   try {
+    const decoded = jwt.verify(token, secret);
+    console.log("Token valido:", decoded);
+
     const itemId = req.params.id;
     const data = req.body;
     const itemRef = db.collection("items").doc(itemId);
     await itemRef.update(data);
     res.status(200).send("item updated");
   } catch (error) {
-    res.status(400).send(error.message);
+    if (error.name === "TokenExpiredError") {
+      // Token ha expirado
+      return res.status(408).send("El token ha expirado");
+    } else {
+      // Otro error
+      return res.status(400).send(error.message);
+    }
   }
 };
 
@@ -152,7 +205,11 @@ exports.deleteItem = async (req, res) => {
       description: '',
       required: true,
     }
-
+    #swagger.parameters['token'] = {
+      in: 'header',
+      description: 'JWT Token',
+      required: true,
+    }
     #swagger.responses[200] = {
       description: '',
     }
@@ -161,11 +218,25 @@ exports.deleteItem = async (req, res) => {
     }
   */
 
+  const token = req.headers["token"];
+  if (!token) {
+    return res.status(400).send("Token no proporcionado");
+  }
+
   try {
+    const decoded = jwt.verify(token, secret);
+    console.log("Token valido:", decoded);
+
     const itemId = req.params.id;
     await db.collection("items").doc(itemId).delete();
     res.status(200).send("Item Detected");
   } catch (error) {
-    res.status(400).send(error.message);
+    if (error.name === "TokenExpiredError") {
+      // Token ha expirado
+      return res.status(408).send("El token ha expirado");
+    } else {
+      // Otro error
+      return res.status(400).send(error.message);
+    }
   }
 };
